pyproject.toml
[tool.poetry]
name = "restaurant-table-reservations"
version = "0.1.0"
description = ""
authors = ["Yana Sozonova <ynakochneva@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"
django = "^5.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"
pillow = "^11.2.1"
ipython = "^9.1.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.1.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '\.git'


[tool.isort]
line_length = 119
